+++ tests/max_len_string_struct1_skyencoder_test_test.go
@@ -160,7 +160,7 @@ func testSkyencoderMaxLenStringStruct1(t *testing.T, obj *MaxLenStringStruct1) {
 	var obj2 MaxLenStringStruct1
 	if n, err := encoder.DeserializeRaw(data1, &obj2); err != nil {
 		t.Fatalf("encoder.DeserializeRaw failed: %v", err)
-	} else if n != len(data1) {
+	} else if n != uint64(len(data1)) {
 		t.Fatalf("encoder.DeserializeRaw failed: %v", encoder.ErrRemainingBytes)
 	}
 	if !cmp.Equal(*obj, obj2, cmpopts.EquateEmpty(), encodertest.IgnoreAllUnexported()) {
