+++ tests/omit_empty_struct_skyencoder_test_test.go
@@ -160,7 +160,7 @@ func testSkyencoderOmitEmptyStruct(t *testing.T, obj *OmitEmptyStruct) {
 	var obj2 OmitEmptyStruct
 	if n, err := encoder.DeserializeRaw(data1, &obj2); err != nil {
 		t.Fatalf("encoder.DeserializeRaw failed: %v", err)
-	} else if n != len(data1) {
+	} else if n != uint64(len(data1)) {
 		t.Fatalf("encoder.DeserializeRaw failed: %v", encoder.ErrRemainingBytes)
 	}
 	if !cmp.Equal(*obj, obj2, cmpopts.EquateEmpty(), encodertest.IgnoreAllUnexported()) {
